Remove: ShiftOS repository
Adding the ShiftOS repository is not useful for devices besides
SHIFTPHONES, while the repo itself is already preconfigured in
the product partition at
`/product/etc/org.fdroid.fdroid/additional_repos.xml`.

Once this functionality is proper in place, the default added
repository is not even useful to users of SHIFT devices,
because the repo itself is maintained by the system and only
causes additional maintenance burden for Neo Store to keep
up-to-date.

For example at the time of writing this commit the public
signature of the repo has changed.

See:
  - https://github.com/SHIFTPHONES/android_vendor_shiftos/blob/08c46cb1eb945603cb51c8985ef70f7ffc65d8db/prebuilt/common/etc/org.fdroid.fdroid/additional_repos.xml

Signed-off-by: Alexander Martinz <amartinz@shiftphones.com>